pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pygentic"
version = "0.1.2"
description = "LLM Mad-libs - Self-generating Pydantic models powered by Mirascope"
license = {file = "LICENSE"}
authors = [{name = "Bullish Design", email = "BullishDesignEngineering@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
    "mirascope[google,anthropic,openai]>=1.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pygentic"]
